Conflicts:
	Projects/Win32Framework/Win32Framework.vcxproj
	Projects/Win32Framework/Win32Framework.vcxproj.filters
	Sources/Internal/Job/JobManager.cpp
	Sources/Internal/Job/JobWaiter.cpp
	Sources/Internal/Job/JobWaiter.h
	Sources/Internal/Render/2D/TextBlock.cpp
	Sources/Internal/Render/2D/TextBlock.h

Former-commit-id: 04f3fac60b0d4aea4952d94f7b4a287047e8d460

+++ Sources/Internal/Render/2D/TextBlockSoftwareRender.h
@@ -64,7 +64,7 @@ protected:
 							  int32 lineSize);
 
 private:
-	int16* buf;
+    int8* buf;
 	FTFont* ftFont;
 };
